requirements.txt
streamlit
pandas
supabase
fpdf
xlsxwriter
streamlit_cookies_controller
plotly.express